First Migration

+++ Sales.Domain/Departments/Department.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Sales.Domain.Departments
 {
-    public partial class Department : BaseEntity<short>
+    public partial class Department : BaseEntity<int>
     {
         public string Name { get; internal set; }
         public string Description { get; internal set; }
